+++ core/src/main/java/com/app/core/data/datasource/WeatherDataSource.kt
@@ -1,7 +1,7 @@
 package com.app.core.data.datasource
 
-import com.app.core.data.repository.ResultWrapper
 import com.app.core.domain.ForecastResponse
+import com.app.core.domain.ResultWrapper
 import kotlinx.coroutines.flow.Flow
 
 interface WeatherDataSource {
