+++ core/src/main/java/com/app/core/data/repository/home/WeatherRepository.kt
@@ -1,7 +1,7 @@
 package com.app.core.data.repository.home
 
-import com.app.core.data.repository.ResultWrapper
 import com.app.core.domain.ForecastResponse
+import com.app.core.domain.ResultWrapper
 import kotlinx.coroutines.flow.Flow
 
 interface WeatherRepository {
