+++ core/src/main/java/com/app/core/domain/ResultWrapper.kt
@@ -1,0 +1,8 @@
+package com.app.core.domain
+
+sealed class ResultWrapper<out T : Any> {
+    data class Success<out T : Any>(val data: T) : ResultWrapper<T>()
+    data class Error(val exception : Exception) : ResultWrapper<Nothing>()
+    data class ErrorString(val exception : String) : ResultWrapper<Nothing>()
+    object InProgress : ResultWrapper<Nothing>()
+}
